# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/operator": {
            "post": {
                "description": "Add an operator to the store",
                "tags": [
                    "Operator"
                ],
                "summary": "Add an operator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operator added",
                        "schema": {
                            "$ref": "#/definitions/src.SimpleJSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/src.SimpleJSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{name}": {
            "get": {
                "description": "Returns json object with operator description",
                "tags": [
                    "Operator"
                ],
                "summary": "Describes an operator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operator description",
                        "schema": {
                            "$ref": "#/definitions/src.OperatorDescription"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{operator}/account": {
            "post": {
                "description": "Add an account with given operator to the store",
                "tags": [
                    "Account"
                ],
                "summary": "Add an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "operator",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account added",
                        "schema": {
                            "$ref": "#/definitions/src.SimpleJSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/src.SimpleJSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{operator}/account/{account}/user": {
            "post": {
                "description": "Add user with given operator and account to the store",
                "tags": [
                    "User"
                ],
                "summary": "Add user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account name",
                        "name": "account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "operator",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User added",
                        "schema": {
                            "$ref": "#/definitions/src.SimpleJSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/src.SimpleJSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{operator}/account/{account}/user/{name}": {
            "get": {
                "description": "Returns json object with user description",
                "tags": [
                    "User"
                ],
                "summary": "Describes user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account name",
                        "name": "account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "operator",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operator description",
                        "schema": {
                            "$ref": "#/definitions/src.AccountDescription"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{operator}/account/{account}/users": {
            "get": {
                "description": "Returns json list of existing users for given operator's account",
                "tags": [
                    "User"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account name",
                        "name": "account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "operator",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users for given operator's account",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{operator}/account/{name}": {
            "get": {
                "description": "Returns json object with account description",
                "tags": [
                    "Account"
                ],
                "summary": "Describes an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "operator",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operator description",
                        "schema": {
                            "$ref": "#/definitions/src.AccountDescription"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{operator}/accounts": {
            "get": {
                "description": "Returns json list of existing accounts for given operator",
                "tags": [
                    "Account"
                ],
                "summary": "List accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator name",
                        "name": "operator",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operator's accounts list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operators": {
            "get": {
                "description": "Returns json list of existing operators",
                "tags": [
                    "Operator"
                ],
                "summary": "List operators",
                "responses": {
                    "200": {
                        "description": "Operators list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "src.AccountDescription": {
            "type": "object",
            "properties": {
                "iat": {
                    "type": "integer"
                },
                "iss": {
                    "type": "string"
                },
                "jti": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nats": {
                    "type": "object",
                    "properties": {
                        "authorization": {
                            "type": "object",
                            "properties": {
                                "auth_users": {}
                            }
                        },
                        "default_permissions": {
                            "type": "object",
                            "properties": {
                                "pub": {
                                    "type": "object",
                                    "additionalProperties": true
                                },
                                "sub": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            }
                        },
                        "limits": {
                            "type": "object",
                            "properties": {
                                "conn": {
                                    "type": "integer"
                                },
                                "data": {
                                    "type": "integer"
                                },
                                "exports": {
                                    "type": "integer"
                                },
                                "imports": {
                                    "type": "integer"
                                },
                                "leaf": {
                                    "type": "integer"
                                },
                                "payload": {
                                    "type": "integer"
                                },
                                "subs": {
                                    "type": "integer"
                                },
                                "wildcards": {
                                    "type": "boolean"
                                }
                            }
                        },
                        "type": {
                            "type": "string"
                        },
                        "version": {
                            "type": "integer"
                        }
                    }
                },
                "sub": {
                    "type": "string"
                }
            }
        },
        "src.OperatorDescription": {
            "type": "object",
            "properties": {
                "iat": {
                    "type": "integer"
                },
                "iss": {
                    "type": "string"
                },
                "jti": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nats": {
                    "type": "object",
                    "properties": {
                        "type": {
                            "type": "string"
                        },
                        "version": {
                            "type": "integer"
                        }
                    }
                },
                "sub": {
                    "type": "string"
                }
            }
        },
        "src.SimpleJSONResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}